fix: added ICommand, IMoving, MoveCommand
