+++ SpaceBattle.Lib/ICommand.cs
@@ -1,0 +1,5 @@
+namespace SpaceBattle.Lib;
+public interface ICommand
+{
+    public void Execute();
+}
