+++ SpaceBattle.Lib/IMoving.cs
@@ -1,0 +1,7 @@
+﻿namespace SpaceBattle.Lib;
+
+public interface IMoving
+{
+    int[] Position { get; set; }
+    int[] Velocity { get; }
+}
