+++ SpaceBattle.Tests/MoveCommandTest.cs
@@ -8,22 +8,14 @@ public class MoveCommandTest
     [Fact]
     public void TestPostitive()
     {
-        // AAA
-        // Arrange
-        //1. Создать движущийся объект
         var moving = new Mock<IMoving>();
 
-        //2. Движущемуся объекту присвоить позицию и скорость
         moving.SetupGet(m => m.Position).Returns(new int[] { 12, 5 });
         moving.SetupGet(m => m.Velocity).Returns(new int[] { -7, 3 });
 
-        // Act
-        // Вызвать команду движения
         var cmd = new MoveCommand(moving.Object);
         cmd.Execute();
 
-        // Assert
-        // Проверить, что объект находится в ожидаемой позиции
         moving.VerifySet(m => m.Position = new int[] { 5, 8 });
     }
 
@@ -36,8 +28,8 @@ public class MoveCommandTest
 
         var cmd = new MoveCommand(moving.Object);
 
-        Assert.Throws<Exception>( // Assert
-            () => cmd.Execute() // Act
+        Assert.Throws<Exception>( 
+            () => cmd.Execute()
         );
 
     }
